# docs/openapi.json
{
  "openapi": "3.0.3",
...
  "info": {
    "title": "Splitt MCP Server",
    "description": "MCP (Model Context Protocol) server for Splitt, the outdoor gear rental marketplace. Exposes role-aware tools for renters and vendors over MCP Streamable HTTP, plus an OAuth 2.1 authorization server so MCP clients can sign users in with their Splitt account.",
    "version": "2.0.0",
    "contact": {
      "name": "Splitt",
      "url": "https://go-splitt.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "https://mcp-server-splitmygear.vercel.app"
    }
  ],
  "components": {
    "securitySchemes": {
      "operatorKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Operator key: public tools only."
      },
      "oauth": {
        "type": "oauth2",
        "description": "Sign in as a Splitt user (PKCE S256, public clients).",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "/oauth/authorize",
            "tokenUrl": "/oauth/token",
            "refreshUrl": "/oauth/token",
            "scopes": {}
          }
        }
      },
      "backendJwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "A raw Splitt backend JWT (first-party integrations)."
      }
    },
    "schemas": {
      "JsonRpcRequest": {
        "type": "object",
        "required": [
          "jsonrpc",
          "method"
        ],
        "properties": {
          "jsonrpc": {
            "type": "string",
            "enum": [
              "2.0"
            ]
          },
          "id": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ]
          },
          "method": {
            "type": "string",
            "example": "tools/call"
          },
          "params": {
            "type": "object"
          }
        }
      },
      "JsonRpcResponse": {
        "type": "object",
        "properties": {
          "jsonrpc": {
            "type": "string"
          },
          "id": {},
          "result": {
            "type": "object"
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "integer"
              },
              "message": {
                "type": "string"
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
    "/api/mcp": {
      "post": {
        "summary": "MCP Streamable HTTP endpoint",
        "description": "JSON-RPC 2.0 over HTTP (stateless; JSON responses). The tool list returned by tools/list depends on the authenticated principal. A 401 carries WWW-Authenticate with the RFC 9728 resource_metadata URL.",
        "security": [
          {
            "operatorKey": []
          },
          {
            "oauth": []
          },
          {
            "backendJwt": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JsonRpcRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "JSON-RPC response",
            "headers": {
              "X-RateLimit-Remaining": {
                "schema": {
                  "type": "integer"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JsonRpcResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "headers": {
              "WWW-Authenticate": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited"
          }
        }
      },
      "get": {
        "summary": "MCP SSE stream (not supported in stateless mode)",
        "responses": {
          "405": {
            "description": "Method not allowed"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "summary": "MCP session termination (no-op in stateless mode)",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "options": {
        "summary": "CORS preflight",
        "responses": {
          "204": {
            "description": "No content"
          }
        }
      }
    },
...
  }
}